move more tests to MockServer
