+++ libraries/apollo-mockserver/api/apollo-mockserver.api
@@ -47,6 +47,7 @@ public abstract interface class com/apollographql/apollo3/mockserver/MockServer 
 	public static synthetic fun awaitAnyRequest-VtjQ1oo$default (Lcom/apollographql/apollo3/mockserver/MockServer;JLkotlin/coroutines/Continuation;ILjava/lang/Object;)Ljava/lang/Object;
 	public abstract fun close ()V
 	public abstract fun enqueue (Lcom/apollographql/apollo3/mockserver/MockResponse;)V
+	public abstract fun port (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun stop (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public static synthetic fun stop$suspendImpl (Lcom/apollographql/apollo3/mockserver/MockServer;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public abstract fun takeRequest ()Lcom/apollographql/apollo3/mockserver/MockRequest;
